Responding to Scott's review     - moved SNFileDBObject to DatabaseEmulators from CatalogClasses     - removed galTileID reference     - Added blank line to Catalog Classes
	modified:   CatalogClasses.py
	modified:   DatabaseEmulator.py
